minor refactor: rename renderer => integration
