+++ packages/vike-react-zustand/package.json
@@ -6,8 +6,8 @@
   "types": "dist/index.d.ts",
   "exports": {
     ".": "./dist/index.js",
-    "./config": "./dist/renderer/+config.js",
-    "./renderer/Wrapper": "./dist/renderer/Wrapper.js",
+    "./config": "./dist/integration/+config.js",
+    "./integration/Wrapper": "./dist/integration/Wrapper.js",
     "./plugin": "./dist/plugin.js"
   },
   "scripts": {
@@ -47,10 +47,10 @@
   "typesVersions": {
     "*": {
       "config": [
-        "dist/renderer/+config.d.ts"
+        "dist/integration/+config.d.ts"
       ],
-      "renderer/Wrapper": [
-        "dist/renderer/Wrapper.d.ts"
+      "integration/Wrapper": [
+        "dist/integration/Wrapper.d.ts"
       ],
       "plugin": [
         "dist/plugin.d.ts"
