+++ packages/vike-react-zustand/src/integration/types.d.ts
