+++ packages/vike-react-zustand/src/plugin.ts
@@ -1,7 +1,7 @@
 export { vikeReactZustand }
 
 import type { Plugin } from 'vite'
-import { initializers_remove } from './renderer/context.js'
+import { initializers_remove } from './integration/context.js'
 import { assert } from './utils.js'
 import { init, parse } from 'es-module-lexer'
